+++ src/clj/clojure/main.clj
@@ -13,7 +13,10 @@
   clojure.main
   (:refer-clojure :exclude [with-bindings])
   (:require [clojure.spec.alpha :as spec])
-  (:import (clojure.lang Compiler Compiler$CompilerException
+  (:import (java.io StringReader BufferedWriter FileWriter)
+           (java.nio.file Files)
+           (java.nio.file.attribute FileAttribute)
+           (clojure.lang Compiler Compiler$CompilerException
                          LineNumberingPushbackReader RT LispReader$ReaderException))
   ;;(:use [clojure.repl :only (demunge root-cause stack-element-str)])
   )
@@ -21,12 +24,6 @@
 (declare main)
 
 ;;;;;;;;;;;;;;;;;;; redundantly copied from clojure.repl to avoid dep ;;;;;;;;;;;;;;
-#_(defn root-cause [x] x)
-#_(defn stack-element-str
-  "Returns a (possibly unmunged) string representation of a StackTraceElement"
-  {:added "1.3"}
-  [^StackTraceElement el]
-  (.getClassName el))
 
 (defn demunge
   "Given a string representation of a fn class,
@@ -47,6 +44,20 @@
       (if-let [cause (.getCause cause)]
         (recur cause)
         cause))))
+
+;;;;;;;;;;;;;;;;;;; end of redundantly copied from clojure.repl to avoid dep ;;;;;;;;;;;;;;
+
+(def ^:private core-namespaces
+  #{"clojure.core" "clojure.core.reducers" "clojure.core.protocols" "clojure.data" "clojure.datafy"
+    "clojure.edn" "clojure.instant" "clojure.java.io" "clojure.main" "clojure.pprint" "clojure.reflect"
+    "clojure.repl" "clojure.set" "clojure.spec.alpha" "clojure.spec.gen.alpha" "clojure.spec.test.alpha"
+    "clojure.string" "clojure.template" "clojure.uuid" "clojure.walk" "clojure.xml" "clojure.zip"})
+
+(defn- core-class?
+  [^String class-name]
+  (and (not (nil? class-name))
+       (or (.startsWith class-name "clojure.lang.")
+           (contains? core-namespaces (second (re-find #"^([^$]+)\$" class-name))))))
 
 (defn stack-element-str
   "Returns a (possibly unmunged) string representation of a StackTraceElement"
@@ -126,6 +137,20 @@
      (or (Character/isWhitespace (char c)) (= c (int \,))) (recur (.read s))
      :else (do (.unread s c) :body))))
 
+(defn renumbering-read
+  "Reads from reader, which must be a LineNumberingPushbackReader, while capturing
+  the read string. If the read is successful, reset the line number and re-read.
+  The line number on re-read is the passed line-number unless :line or
+  :clojure.core/eval-file meta are explicitly set on the read value."
+  {:added "1.10"}
+  ([opts ^LineNumberingPushbackReader reader line-number]
+   (let [pre-line (.getLineNumber reader)
+         [pre-read s] (read+string opts reader)
+         {:keys [clojure.core/eval-file line]} (meta pre-read)
+         re-reader (doto (LineNumberingPushbackReader. (StringReader. s))
+                     (.setLineNumber (if (and line (or eval-file (not= pre-line line))) line line-number)))]
+     (read opts re-reader))))
+
 (defn repl-read
   "Default :read hook for repl. Reads from *in* which must either be an
   instance of LineNumberingPushbackReader or duplicate its behavior of both
@@ -140,7 +165,7 @@
   [request-prompt request-exit]
   (or ({:line-start request-prompt :stream-end request-exit}
        (skip-whitespace *in*))
-      (let [input (read {:read-cond :allow} *in*)]
+      (let [input (renumbering-read {:read-cond :allow} *in* 1)]
         (skip-if-eol *in*)
         input)))
 
@@ -149,72 +174,182 @@
   [throwable]
   (root-cause throwable))
 
-(defn- init-cause
-  "Returns initial root cause exception (deepest cause in the exception chain)."
-  ^Throwable [t]
-  (loop [^Throwable cause t]
-    (if-let [cause (.getCause cause)]
-      (recur cause)
-      cause)))
+(defn- file-name
+  "Helper to get just the file name part of a path or nil"
+  [^String full-path]
+  (when full-path
+    (try
+      (.getName (java.io.File. full-path))
+      (catch Throwable t))))
+
+(defn- file-path
+  "Helper to get the relative path to the source file or nil"
+  [^String full-path]
+  (when full-path
+    (try
+      (let [path (.getPath (java.io.File. full-path))
+            cd-path (str (.getAbsolutePath (java.io.File. "")) "/")]
+        (if (.startsWith path cd-path)
+          (subs path (count cd-path))
+          path))
+      (catch Throwable t
+        full-path))))
+
+(defn- java-loc->source
+  "Convert Java class name and method symbol to source symbol, either a
+  Clojure function or Java class and method."
+  [clazz method]
+  (if (#{'invoke 'invokeStatic} method)
+    (let [degen #(.replaceAll ^String % "--.*$" "")
+          [ns-name fn-name & nested] (->> (str clazz) (.split #"\$") (map demunge) (map degen))]
+      (symbol ns-name (String/join "$" ^"[Ljava.lang.String;" (into-array String (cons fn-name nested)))))
+    (symbol (name clazz) (name method))))
+
+(defn ex-triage
+  "Returns an analysis of the phase, error, cause, and location of an error that occurred
+  based on Throwable data, as returned by Throwable->map. All attributes other than phase
+  are optional:
+    :clojure.error/phase - keyword phase indicator, one of:
+      :read-source :compile-syntax-check :compilation :macro-syntax-check :macroexpansion
+      :execution :read-eval-result :print-eval-result
+    :clojure.error/source - file name (no path)
+    :clojure.error/path - source path
+    :clojure.error/line - integer line number
+    :clojure.error/column - integer column number
+    :clojure.error/symbol - symbol being expanded/compiled/invoked
+    :clojure.error/class - cause exception class symbol
+    :clojure.error/cause - cause exception message
+    :clojure.error/spec - explain-data for spec error"
+  {:added "1.10"}
+  [datafied-throwable]
+  (let [{:keys [via trace phase] :or {phase :execution}} datafied-throwable
+        {:keys [type message data]} (last via)
+        {:clojure.spec.alpha/keys [problems fn], :clojure.spec.test.alpha/keys [caller]} data
+        {:clojure.error/keys [source] :as top-data} (:data (first via))]
+    (assoc
+      (case phase
+        :read-source
+        (let [{:clojure.error/keys [line column]} data]
+          (cond-> (merge (-> via second :data) top-data)
+            source (assoc :clojure.error/source (file-name source)
+                          :clojure.error/path (file-path source))
+            (#{"NO_SOURCE_FILE" "NO_SOURCE_PATH"} source) (dissoc :clojure.error/source :clojure.error/path)
+            message (assoc :clojure.error/cause message)))
+
+        (:compile-syntax-check :compilation :macro-syntax-check :macroexpansion)
+        (cond-> top-data
+          source (assoc :clojure.error/source (file-name source)
+                        :clojure.error/path (file-path source))
+          (#{"NO_SOURCE_FILE" "NO_SOURCE_PATH"} source) (dissoc :clojure.error/source :clojure.error/path)
+          type (assoc :clojure.error/class type)
+          message (assoc :clojure.error/cause message)
+          problems (assoc :clojure.error/spec data))
+
+        (:read-eval-result :print-eval-result)
+        (let [[source method file line] (-> trace first)]
+          (cond-> top-data
+            line (assoc :clojure.error/line line)
+            file (assoc :clojure.error/source file)
+            (and source method) (assoc :clojure.error/symbol (java-loc->source source method))
+            type (assoc :clojure.error/class type)
+            message (assoc :clojure.error/cause message)))
+
+        :execution
+        (let [[source method file line] (->> trace (drop-while #(core-class? (name (first %)))) first)
+              file (first (remove #(or (nil? %) (#{"NO_SOURCE_FILE" "NO_SOURCE_PATH"} %)) [(:file caller) file]))
+              err-line (or (:line caller) line)]
+          (cond-> {:clojure.error/class type}
+            err-line (assoc :clojure.error/line err-line)
+            message (assoc :clojure.error/cause message)
+            (or fn (and source method)) (assoc :clojure.error/symbol (or fn (java-loc->source source method)))
+            file (assoc :clojure.error/source file)
+            problems (assoc :clojure.error/spec data))))
+      :clojure.error/phase phase)))
 
 (defn ex-str
-  "Returns a string from an exception for printing at the repl.
-  The first line summarizes data from the exception instance: phase, location,
-  cause message, etc. The subsequent lines contain ex-data info if available."
+  "Returns a string from exception data, as produced by ex-triage.
+  The first line summarizes the exception phase and location.
+  The subsequent lines describe the cause."
+  {:added "1.10"}
+  [{:clojure.error/keys [phase source path line column symbol class cause spec]
+    :as triage-data}]
+  (let [loc (str (or path source "REPL") ":" (or line 1) (if column (str ":" column) ""))
+        class-name (name (or class ""))
+        simple-class (if class (or (first (re-find #"([^.])++$" class-name)) class-name))
+        cause-type (if (contains? #{"Exception" "RuntimeException"} simple-class)
+                     "" ;; omit, not useful
+                     (str " (" simple-class ")"))]
+    (case phase
+      :read-source
+      (format "Syntax error reading source at (%s).%n%s%n" loc cause)
+
+      :macro-syntax-check
+      (format "Syntax error macroexpanding %sat (%s).%n%s"
+              (if symbol (str symbol " ") "")
+              loc
+              (if spec
+                (with-out-str
+                  (spec/explain-out
+                    (if (= spec/*explain-out* spec/explain-printer)
+                      (update spec :clojure.spec.alpha/problems
+                              (fn [probs] (map #(dissoc % :in) probs)))
+                      spec)))
+                (format "%s%n" cause)))
+
+      :macroexpansion
+      (format "Unexpected error%s macroexpanding %sat (%s).%n%s%n"
+              cause-type
+              (if symbol (str symbol " ") "")
+              loc
+              cause)
+
+      :compile-syntax-check
+      (format "Syntax error%s compiling %sat (%s).%n%s%n"
+              cause-type
+              (if symbol (str symbol " ") "")
+              loc
+              cause)
+
+      :compilation
+      (format "Unexpected error%s compiling %sat (%s).%n%s%n"
+              cause-type
+              (if symbol (str symbol " ") "")
+              loc
+              cause)
+
+      :read-eval-result
+      (format "Error reading eval result%s at %s (%s).%n%s%n" cause-type symbol loc cause)
+
+      :print-eval-result
+      (format "Error printing return value%s at %s (%s).%n%s%n" cause-type symbol loc cause)
+
+      :execution
+      (if spec
+        (format "Execution error - invalid arguments to %s at (%s).%n%s"
+                symbol
+                loc
+                (with-out-str
+                  (spec/explain-out
+                    (if (= spec/*explain-out* spec/explain-printer)
+                      (update spec :clojure.spec.alpha/problems
+                              (fn [probs] (map #(dissoc % :in) probs)))
+                      spec))))
+        (format "Execution error%s at %s(%s).%n%s%n"
+                cause-type
+                (if symbol (str symbol " ") "")
+                loc
+                cause)))))
+
+(defn err->msg
+  "Helper to return an error message string from an exception."
   [^Throwable e]
-  (let [msg (.getMessage e)
-        cause (init-cause e)
-        tr (.getStackTrace cause)
-        el (when-not (zero? (count tr)) (aget tr 0))
-        st (if el (stack-element-str el) "[trace missing]") ;; jvm may omit stack
-        top-data (ex-data e)
-        data (ex-data cause)]
-    (str
-      (case (:clojure.error/phase top-data)
-        :read
-        (if (instance? Compiler$CompilerException e)
-          (.toString e)
-          (format "%s.%n%s" (.getMessage e) (.getMessage (init-cause e))))
-
-        (:compile :macroexpand)
-        (.toString e)
-
-        :print
-        (format "Error printing return value (%s) at %s.%n%s"
-          (.. cause getClass getSimpleName)
-          st
-          (.getMessage cause))
-
-        ;; eval
-        (if (and data (contains? data :clojure.spec.alpha/problems))
-          (let [{:clojure.spec.alpha/keys [fn]} data]
-            (format "Evaluation error - invalid arguments to %s at %s."
-              fn
-              (if-let [{:keys [file line]} (:clojure.spec.test.alpha/caller data)]
-                (format "(%s:%s)" file line)
-                st)))
-          (format "Evaluation error (%s) at %s.%n%s"
-            (.. cause getClass getSimpleName)
-            st
-            (.getMessage cause))))
-
-      (if data
-        (if (contains? data :clojure.spec.alpha/problems)
-          (format "%n%s"
-            (with-out-str
-              (spec/explain-out
-                (if (= spec/*explain-out* spec/explain-printer)
-                  (update-in data [:clojure.spec.alpha/problems]
-                    (fn [probs] (map #(dissoc % :in) probs)))
-                  data))))
-          (System/lineSeparator))
-        (System/lineSeparator)))))
+  (-> e Throwable->map ex-triage ex-str))
 
 (defn repl-caught
   "Default :caught hook for repl"
   [e]
   (binding [*out* *err*]
-    (print (ex-str e))
+    (print (err->msg e))
     (flush)))
 
 (def ^{:doc "A sequence of lib specs that are applied to `require`
@@ -297,12 +432,7 @@ by default when a new command-line REPL is started."} repl-requires
                   input (try
                           (with-read-known (read request-prompt request-exit))
                           (catch LispReader$ReaderException e
-                            (throw (ex-info
-                                     (str "Syntax error reading source at (" (.-line e) ":" (.-column e) ")")
-                                     {:clojure.error/phase :read
-                                      :clojure.error/line (.-line e)
-                                      :clojure.error/column (.-column e)}
-                                     e))))]
+                            (throw (ex-info nil {:clojure.error/phase :read-source} e))))]
              (or (#{request-prompt request-exit} input)
                  (let [value (binding [*read-eval* read-eval] (eval input))]
                    (set! *3 *2)
@@ -311,9 +441,7 @@ by default when a new command-line REPL is started."} repl-requires
                    (try
                      (print value)
                      (catch Throwable e
-                       (throw (ex-info (format "Error printing return value. Cause: " (.getMessage e))
-                                {:clojure.error/phase :print}
-                                e)))))))
+                       (throw (ex-info nil {:clojure.error/phase :print-eval-result} e)))))))
            (catch Throwable e
              (caught e)
              (set! *e e))))]
@@ -453,6 +581,38 @@ java -cp clojure.jar clojure.main -i init.clj script.clj args...")
   (let [[inits [sep & args]] (split-with (complement #{"--"}) args)]
     (null-opt args (map vector (repeat "-i") inits))))
 
+(defn report-error
+  "Create and output an exception report for a Throwable to target.
+
+  Options:
+    :target - \"file\" (default), \"stderr\", \"none\"
+
+  If file is specified but cannot be written, falls back to stderr."
+  [^Throwable t & {:keys [target]
+                   :or {target "file"} :as opts}]
+  (when-not (= target "none")
+    (let [trace (Throwable->map t)
+          triage (ex-triage trace)
+          message (ex-str triage)
+          report (array-map
+                   :clojure.main/message message
+                   :clojure.main/triage triage
+                   :clojure.main/trace trace)
+          report-str (with-out-str
+                       (binding [*print-namespace-maps* false]
+                         ((requiring-resolve 'clojure.pprint/pprint) report)))
+          err-path (when (= target "file")
+                     (try
+                       (let [f (.toFile (Files/createTempFile "clojure-" ".edn" (into-array FileAttribute [])))]
+                         (with-open [w (BufferedWriter. (FileWriter. f))]
+                           (binding [*out* w] (println report-str)))
+                         (.getAbsolutePath f))
+                       (catch Throwable _)))] ;; ignore, fallback to stderr
+      (binding [*out* *err*]
+        (if err-path
+          (println (str message (System/lineSeparator) "Full report at:" (System/lineSeparator) err-path))
+          (println (str report-str (System/lineSeparator) message)))))))
+
 (defn main
   "Usage: java -cp clojure.jar clojure.main [init-opt*] [main-opt] [arg*]
 
@@ -461,6 +621,8 @@ java -cp clojure.jar clojure.main -i init.clj script.clj args...")
   init options:
     -i, --init path     Load a file or resource
     -e, --eval string   Evaluate expressions in string; print non-nil values
+    --report target     Report uncaught exception to \"file\" (default), \"stderr\",
+                        or \"none\", overrides System property clojure.main.report
 
   main options:
     -m, --main ns-name  Call the -main function from a namespace with args
@@ -487,11 +649,27 @@ java -cp clojure.jar clojure.main -i init.clj script.clj args...")
   [& args]
   (try
    (if args
-     (loop [[opt arg & more :as args] args inits []]
-       (if (init-dispatch opt)
-         (recur more (conj inits [opt arg]))
-         ((main-dispatch opt) args inits)))
-     (repl-opt nil nil))
+     (loop [[opt arg & more :as args] args, inits [], flags nil]
+       (cond
+         ;; flag
+         (contains? #{"--report"} opt)
+         (recur more inits (merge flags {(subs opt 2) arg}))
+
+         ;; init opt
+         (init-dispatch opt)
+         (recur more (conj inits [opt arg]) flags)
+
+         :main-opt
+         (try
+           ((main-dispatch opt) args inits)
+           (catch Throwable t
+             (report-error t :target (get flags "report" (System/getProperty "clojure.main.report" "file")))
+             (System/exit 1)))))
+     (try
+       (repl-opt nil nil)
+       (catch Throwable t
+         (report-error t :target "file")
+         (System/exit 1))))
    (finally 
      (flush))))