+++ src/jvm/clojure/lang/Compiler.java
@@ -474,7 +474,7 @@ static class DefExpr implements Expr{
 		catch(Throwable e)
 			{
 			if(!(e instanceof CompilerException))
-				throw new CompilerException(source, line, column, Compiler.DEF, e);
+				throw new CompilerException(source, line, column, Compiler.DEF, CompilerException.PHASE_EXECUTION, e);
 			else
 				throw (CompilerException) e;
 			}
@@ -1602,7 +1602,7 @@ static class InstanceMethodExpr extends MethodExpr{
 		catch(Throwable e)
 			{
 			if(!(e instanceof CompilerException))
-				throw new CompilerException(source, line, column, e);
+				throw new CompilerException(source, line, column, null, CompilerException.PHASE_EXECUTION, e);
 			else
 				throw (CompilerException) e;
 			}
@@ -1782,7 +1782,7 @@ static class StaticMethodExpr extends MethodExpr{
 		catch(Throwable e)
 			{
 			if(!(e instanceof CompilerException))
-				throw new CompilerException(source, line, column, e);
+				throw new CompilerException(source, line, column, null, CompilerException.PHASE_EXECUTION, e);
 			else
 				throw (CompilerException) e;
 			}
@@ -3357,7 +3357,7 @@ static class KeywordInvokeExpr implements Expr{
 		catch(Throwable e)
 			{
 			if(!(e instanceof CompilerException))
-				throw new CompilerException(source, line, column, e);
+				throw new CompilerException(source, line, column, null, CompilerException.PHASE_EXECUTION, e);
 			else
 				throw (CompilerException) e;
 			}
@@ -3749,7 +3749,7 @@ static class InvokeExpr implements Expr{
 		catch(Throwable e)
 			{
 			if(!(e instanceof CompilerException))
-				throw new CompilerException(source, line, column, e);
+				throw new CompilerException(source, line, column, null, CompilerException.PHASE_EXECUTION, e);
 			else
 				throw (CompilerException) e;
 			}
@@ -6898,16 +6898,19 @@ static public class CompilerException extends RuntimeException implements IExcep
 
     // Error keys
     final static public String ERR_NS = "clojure.error";
-    final static public Keyword ERR_SOURCE = Keyword.intern(ERR_NS, "source"); // :clojure.error/source
-    final static public Keyword ERR_LINE = Keyword.intern(ERR_NS, "line");     // :clojure.error/line
-    final static public Keyword ERR_COLUMN = Keyword.intern(ERR_NS, "column"); // :clojure.error/column
-    final static public Keyword ERR_PHASE = Keyword.intern(ERR_NS, "phase");   // :clojure.error/phase
-    final static public Keyword ERR_SYMBOL = Keyword.intern(ERR_NS, "symbol"); // :clojure.error/symbol
+    final static public Keyword ERR_SOURCE = Keyword.intern(ERR_NS, "source");
+    final static public Keyword ERR_LINE = Keyword.intern(ERR_NS, "line");
+    final static public Keyword ERR_COLUMN = Keyword.intern(ERR_NS, "column");
+    final static public Keyword ERR_PHASE = Keyword.intern(ERR_NS, "phase");
+    final static public Keyword ERR_SYMBOL = Keyword.intern(ERR_NS, "symbol");
 
     // Compile error phases
-    final static public Keyword PHASE_READ = Keyword.intern(null, "read");                // :read
-    final static public Keyword PHASE_MACROEXPAND = Keyword.intern(null, "macroexpand");  // :macroexpand
-    final static public Keyword PHASE_COMPILE = Keyword.intern(null, "compile");          // :compile
+    final static public Keyword PHASE_READ = Keyword.intern(null, "read-source");
+    final static public Keyword PHASE_MACRO_SYNTAX_CHECK = Keyword.intern(null, "macro-syntax-check");
+    final static public Keyword PHASE_MACROEXPANSION = Keyword.intern(null, "macroexpansion");
+    final static public Keyword PHASE_COMPILE_SYNTAX_CHECK = Keyword.intern(null, "compile-syntax-check");
+    final static public Keyword PHASE_COMPILATION = Keyword.intern(null, "compilation");
+    final static public Keyword PHASE_EXECUTION = Keyword.intern(null, "execution");
 
 	final static public Keyword SPEC_PROBLEMS = Keyword.intern("clojure.spec.alpha", "problems");
 
@@ -6917,14 +6920,14 @@ static public class CompilerException extends RuntimeException implements IExcep
 	}
 
 	public CompilerException(String source, int line, int column, Symbol sym, Throwable cause){
-		this(source, line, column, sym, PHASE_COMPILE, cause);
+		this(source, line, column, sym, PHASE_COMPILE_SYNTAX_CHECK, cause);
 	}
 
 	public CompilerException(String source, int line, int column, Symbol sym, Keyword phase, Throwable cause){
 		super(makeMsg(source, line, column, sym, phase, cause), cause);
 		this.source = source;
 		this.line = line;
-		Associative m = RT.map(ERR_LINE, line, ERR_COLUMN, column, ERR_PHASE, phase);
+		Associative m = RT.map(ERR_PHASE, phase, ERR_LINE, line, ERR_COLUMN, column);
 		if(source != null) m = RT.assoc(m, ERR_SOURCE, source);
 		if(sym != null) m = RT.assoc(m, ERR_SYMBOL, sym);
 		this.data = m;
@@ -6935,48 +6938,35 @@ static public class CompilerException extends RuntimeException implements IExcep
 	}
 
 	private static String verb(Keyword phase) {
-		if(PHASE_COMPILE.equals(phase)){
-			return "compiling";
-		} else if(PHASE_READ.equals(phase)){
+		if(PHASE_READ.equals(phase)){
 			return "reading source";
+		} else if(PHASE_COMPILE_SYNTAX_CHECK.equals(phase) || PHASE_COMPILATION.equals(phase)){
+			return "compiling";
 		} else {
 			return "macroexpanding";
 		}
 	}
 
-	private static boolean isMacroSyntaxCheck(Throwable e) {
-		return e instanceof IllegalArgumentException ||
-				e instanceof IllegalStateException ||
-				e instanceof ExceptionInfo ||
-				e.getClass().equals(Exception.class);
-	}
-
 	public static String makeMsg(String source, int line, int column, Symbol sym, Keyword phase, Throwable cause){
-		return (phase == PHASE_MACROEXPAND && ! isMacroSyntaxCheck(cause) ? "Unexpected error " : "Syntax error ") +
+		return (PHASE_MACROEXPANSION.equals(phase) ? "Unexpected error " : "Syntax error ") +
 				verb(phase) + " " + (sym != null ? sym + " " : "") +
 				"at (" + (source != null && !source.equals("NO_SOURCE_PATH") ? (source + ":") : "") +
 				line + ":" + column + ").";
 	}
 
-	private static boolean isSpecError(Throwable t) {
-		return (t instanceof IExceptionInfo) && RT.get(((IExceptionInfo) t).getData(), SPEC_PROBLEMS) != null;
-	}
-
 	public String toString(){
 		Throwable cause = getCause();
 		if(cause != null) {
-			if(RT.get(data, ERR_PHASE) == PHASE_MACROEXPAND) {
-				if(isSpecError(cause)) {
+			if (cause instanceof IExceptionInfo) {
+				IPersistentMap data = (IPersistentMap)((IExceptionInfo)cause).getData();
+				if(PHASE_MACRO_SYNTAX_CHECK.equals(data.valAt(ERR_PHASE)) && data.valAt(SPEC_PROBLEMS) != null) {
 					return String.format("%s", getMessage());
 				} else {
 					return String.format("%s%n%s", getMessage(), cause.getMessage());
 				}
-			} else {
-				return String.format("%s%n%s", getMessage(), cause.getMessage());
 			}
-		} else {
-			return getMessage();
 		}
+		return getMessage();
 	}
 }
 
@@ -7057,7 +7047,7 @@ public static void checkSpecs(Var v, ISeq form) {
 		try {
 			ensureMacroCheck().applyTo(RT.cons(v, RT.list(form.next())));
 		} catch(Exception e) {
-			throw new CompilerException((String) SOURCE_PATH.deref(), lineDeref(), columnDeref(), v.toSymbol(), CompilerException.PHASE_MACROEXPAND, e);
+			throw new CompilerException((String) SOURCE_PATH.deref(), lineDeref(), columnDeref(), v.toSymbol(), CompilerException.PHASE_MACRO_SYNTAX_CHECK, e);
 		}
 	}
 }
@@ -7089,16 +7079,23 @@ public static Object macroexpand1(Object x) {
 							throw e;
 						}
 					}
+				catch(CompilerException e)
+					{
+						throw e;
+					}
+				catch(IllegalArgumentException | IllegalStateException | ExceptionInfo e)
+					{
+						throw new CompilerException((String) SOURCE_PATH.deref(), lineDeref(), columnDeref(),
+								(op instanceof Symbol ? (Symbol) op : null),
+								CompilerException.PHASE_MACRO_SYNTAX_CHECK,
+								e);
+					}
 				catch(Throwable e)
 				    {
-						if(e instanceof CompilerException) {
-							throw (CompilerException)e;
-						} else {
-							throw new CompilerException((String) SOURCE_PATH.deref(), lineDeref(), columnDeref(),
-									(op instanceof Symbol ? (Symbol) op : null),
-									CompilerException.PHASE_MACROEXPAND,
-									e);
-						}
+						throw new CompilerException((String) SOURCE_PATH.deref(), lineDeref(), columnDeref(),
+								(op instanceof Symbol ? (Symbol) op : null),
+								(e.getClass().equals(Exception.class) ? CompilerException.PHASE_MACRO_SYNTAX_CHECK : CompilerException.PHASE_MACROEXPANSION),
+								e);
 					}
 			} else
 			{
@@ -7232,13 +7229,21 @@ public static Object eval(Object form, boolean freshLoader) {
 		}
 	try
 		{
-		Object line = lineDeref();
-		Object column = columnDeref();
-		if(RT.meta(form) != null && RT.meta(form).containsKey(RT.LINE_KEY))
-			line = RT.meta(form).valAt(RT.LINE_KEY);
-		if(RT.meta(form) != null && RT.meta(form).containsKey(RT.COLUMN_KEY))
-			column = RT.meta(form).valAt(RT.COLUMN_KEY);
-		Var.pushThreadBindings(RT.map(LINE, line, COLUMN, column));
+		IPersistentMap meta = RT.meta(form);
+		Object line = (meta != null ? meta.valAt(RT.LINE_KEY, LINE.deref()) : LINE.deref());
+		Object column = (meta != null ? meta.valAt(RT.COLUMN_KEY, COLUMN.deref()) : COLUMN.deref());
+		IPersistentMap bindings = RT.mapUniqueKeys(LINE, line, COLUMN, column);
+		if(meta != null) {
+			Object eval_file = meta.valAt(RT.EVAL_FILE_KEY);
+			if(eval_file != null) {
+				bindings = bindings.assoc(SOURCE_PATH, eval_file);
+				try {
+					bindings = bindings.assoc(SOURCE, new File((String)eval_file).getName());
+				} catch (Throwable t) {
+				}
+			}
+		}
+		Var.pushThreadBindings(bindings);
 		try
 			{
 			form = macroexpand(form);