+++ src/jvm/clojure/lang/RT.java
@@ -177,6 +177,7 @@ static Object readTrueFalseUnknown(String s){
 	return Keyword.intern(null, "unknown");
 }
 
+static public final Object REQUIRE_LOCK = new Object();
 static public final Namespace CLOJURE_NS = Namespace.findOrCreate(Symbol.intern("clojure.core"));
 //static final Namespace USER_NS = Namespace.findOrCreate(Symbol.intern("user"));
 final static public Var OUT =
@@ -198,6 +199,7 @@ final static public Var DEFAULT_DATA_READERS = Var.intern(CLOJURE_NS, Symbol.int
 final static public Var SUPPRESS_READ = Var.intern(CLOJURE_NS, Symbol.intern("*suppress-read*"), null).setDynamic();
 final static public Var ASSERT = Var.intern(CLOJURE_NS, Symbol.intern("*assert*"), T).setDynamic();
 final static public Var MATH_CONTEXT = Var.intern(CLOJURE_NS, Symbol.intern("*math-context*"), null).setDynamic();
+static Keyword EVAL_FILE_KEY = Keyword.intern("clojure.core", "eval-file");
 static Keyword LINE_KEY = Keyword.intern(null, "line");
 static Keyword COLUMN_KEY = Keyword.intern(null, "column");
 static Keyword FILE_KEY = Keyword.intern(null, "file");
@@ -333,7 +335,7 @@ static{
 	v.setMeta(map(DOC_KEY, "Sequentially read and evaluate the set of forms contained in the file.",
 	              arglistskw, list(vector(namesym))));
 	try {
-		doInit();
+		load("clojure/core");
 	}
 	catch(Exception e) {
 		throw Util.sneakyThrow(e);
@@ -385,10 +387,6 @@ public static void loadResourceScript(Class c, String name, boolean failIfNotFou
 	else if(failIfNotFound) {
 		throw new FileNotFoundException("Could not locate Clojure resource on classpath: " + name);
 	}
-}
-
-static public void init() {
-	RT.errPrintWriter().println("No need to call RT.init() anymore");
 }
 
 static public long lastModified(URL url, String libfile) throws IOException{
@@ -479,8 +477,13 @@ static public void load(String scriptbase, boolean failIfNotFound) throws IOExce
 			scriptbase.contains("_") ? " Please check that namespaces with dashes use underscores in the Clojure file name." : ""));
 }
 
-static void doInit() throws ClassNotFoundException, IOException{
-	load("clojure/core");
+static public void init() {
+	doInit();
+}
+
+private static boolean INIT = false; // init guard
+private synchronized static void doInit() {
+	if(INIT) {return;} else {INIT=true;}
 
 	Var.pushThreadBindings(
 			RT.mapUniqueKeys(CURRENT_NS, CURRENT_NS.deref(),
@@ -502,6 +505,9 @@ static void doInit() throws ClassNotFoundException, IOException{
 		require.invoke(SERVER);
 		Var start_servers = var("clojure.core.server", "start-servers");
 		start_servers.invoke(System.getProperties());
+	}
+	catch(Exception e) {
+		throw Util.sneakyThrow(e);
 	}
 	finally {
 		Var.popThreadBindings();